Remote Diagnostics Tool (RDT) for IT Admins
Desktop Administrators and Helpdesk staff will benefit from using this Microsoft Windows - Remote Diagnostics Tool.
This remote diagnostics tool (RDT) was designed to aid in the "diagnostics" and "investigation" of a remote PC whilst it is on a network.
No Installation Required! Run on any PC or Server with DOT.NET 4 installed (Windows 7/8/10/11 , Server 2008/2012/2016/2019/22/25).
PLease visit - https://appslife-rdt.appspot.com/ for screen shots of the application in use!